refactor: move Dropdown.tsx to utils folder
